Refactor AtLevel structures
